Utdanning komponent over på felles redux connect
