+++ web/src/frontend/src/digisos/skjema/arbeidUtdanning/utdanning/utdanningTypes.ts
@@ -1,0 +1,9 @@
+export interface Utdanning {
+	"erStudent": null | boolean;
+	"studengradErHeltid": null | boolean;
+}
+
+export const initialUtdanningState: Utdanning = {
+	"erStudent": null,
+	"studengradErHeltid": null
+};
